additional 64 bit cleanup
git-svn-id: http://svn.cip4.org/svnroot/trunk@5603 25ee8a6d-831b-0410-907a-ff8528d65d49

+++ src/jdf/lang/WString.h
@@ -232,8 +232,8 @@ namespace JDF
 		//@}
 
 	public:
-		typedef  unsigned int size_type;
-		static const int npos;
+		typedef size_t size_type;
+		static const size_t npos;
 		typedef WStringIterator iterator;                              // RN20060320
 		typedef const WStringIterator const_iterator;                  // RN20060320
 		typedef JDFCh& reference;
